refactor(io): extract createBaseRouter from PluvServer
